set protected route to delete and create

+++ controllers/projectsController.py
@@ -13,14 +13,29 @@ router = APIRouter()
 views = project_config.views
 
 
+# PROTECTED ROUTE - POST /projects
+
 @router.post("/")
 async def create_project(
-        request: Request, project: modelProject.ProjectSchema):
-    project_created = modelProject.Project.create_project(db, project)
-    if project_created is not False:
-        return {"success": True, "projectCreated": project_created.title}
+        request: Request, project: modelProject.ProjectSchema, authorization: str = Header(None)):
+
+    if authorization == None:
+        print("NO AUTH")
+        return views.TemplateResponse("router/login.html", context={"request": request})
+
+    if verify_token(authorization):
+
+        print("AUTHORIZED")
+
+        project_created = modelProject.Project.create_project(db, project)
+
+        if project_created is not False:
+            return {"success": True, "projectCreated": project_created.title}
+        else:
+            return {"success": False, "message": "failed creating new project"}
+
     else:
-        return {"success": False, "message": "failed creating new project"}
+        return {"sucess": False, "message": "Not AAuthorized"}
 
 
 @router.get("/")
@@ -31,6 +46,8 @@ async def get_all_projects(request: Request):
             "request": request, "projects": projects})
 
 
+# PROTECTED ROUTE - return the index view with administrator button like edit or delete - GET /projects/admin-index
+
 @router.get("/admin-index")
 async def get_all_projects_admin(request: Request, authorization: str = Header(None)):
 
@@ -38,41 +55,39 @@ async def get_all_projects_admin(request: Request, authorization: str = Header(N
         print("NO AUTH")
         return views.TemplateResponse("router/login.html", context={"request": request})
 
-    try:
-        if verify_token(authorization):
+    if verify_token(authorization):
 
-            print("AUTHORIZED")
+        print("AUTHORIZED")
 
-            projects = modelProject.Project.get_all_projects(db)
-            return views.TemplateResponse(
-                "project/index.html", status_code=200, context={"request": request,
-                                                                "projects": projects,
-                                                                "admin": True})
+        projects = modelProject.Project.get_all_projects(db)
+        return views.TemplateResponse(
+            "project/index.html", status_code=200, context={"request": request,
+                                                            "projects": projects,
+                                                            "admin": True})
+    else:
+        return {"success": False, "error": "Not Authorized"}
 
-    except Exception as err:
 
-        print("NOT VALID TOKEN")
-
-        return {"success": False, "error": err}
-
-        # return views.TemplateResponse(
-        #     "project/index.html",
-        #     context={"request": request, "projects": projects, "admin": True})
-
-# DELETE - delete the project - GET /projects/{id}/delete
+# PROTECTED ROUTE - DELETE - delete the project - GET /projects/{id}/delete
 
 
 @router.get("/{project_id}/delete")
-async def delete_project(request: Request, project_id: int):
-    if modelProject.Project.delete_project(db, project_id):
-        projects = modelProject.Project.get_all_projects(db)
-        return views.TemplateResponse(
-            "project/index.html",
-            context={"request": request, 'projects': projects})
-    else:
-        return views.TemplateResponse(
-            "project/index.html",
-            context={"request": request, 'flashMessage': "an error occurred"})
+async def delete_project(request: Request, project_id: int, authorization: str = Header(None)):
+
+    if authorization == None:
+        print("NO AUTH")
+        return views.TemplateResponse("router/login.html", context={"request": request})
+
+    if verify_token(authorization):
+        if modelProject.Project.delete_project(db, project_id):
+            projects = modelProject.Project.get_all_projects(db)
+            return views.TemplateResponse(
+                "project/index.html",
+                context={"request": request, 'projects': projects})
+        else:
+            return views.TemplateResponse(
+                "project/index.html",
+                context={"request": request, 'flashMessage': "an error occurred"})
 
 # EDIT - get the edit form - GET /projects/{id}/edit
